refactor(plugins): rename plugins directory
