+++ .env.local.example
@@ -1,15 +1,19 @@
-# RPC urls, follow the format RPC_URL_{chainId}={rpcUrl}
+# RPC configuration
+# Follow the format: RPC_URL_{chainId}={rpcUrl}
 
-PONDER_RPC_URL_1=https://ethereum-rpc.publicnode.com
-PONDER_RPC_URL_8453=https://base-rpc.publicnode.com
-PONDER_RPC_URL_59144=https://linea-rpc.publicnode.com
+RPC_URL_1=https://ethereum-rpc.publicnode.com
+RPC_URL_8453=https://base-rpc.publicnode.com
+RPC_URL_59144=https://linea-rpc.publicnode.com
 
-# ponder indexer ens deployment configuration
+# Identify which indexer plugin to activate (see `src/plugins` for available plugins)
 
-INDEX_SUBNAME=base.eth
+ACTIVE_PLUGIN=base.eth
 
-# ponder indexer database configuration
+# Database configuration
 
-## one schema name per chain ID, i.e. ponder_ens_${INDEX_SUBNAME}
-DATABASE_SCHEMA=ponder_ens_base.eth
+# This is where the indexer will create the tables defined in ponder.schema.ts
+# No two indexer instances can use the same database schema at the same time. This prevents data corruption.
+# @link https://ponder.sh/docs/api-reference/database#database-schema-rules
+DATABASE_SCHEMA=subname_index_base.eth
+# The indexer will use Postgres with that as the connection string. If not defined, the indexer will use PSlite.
 DATABASE_URL=postgresql://dbuser:abcd1234@localhost:5432/my_database
