+++ src/plugins/README.md
@@ -1,0 +1,8 @@
+# Indexer plugins
+
+This directory contains plugins which allow defining subname-specific processing of blockchain events.
+Only one plugin can be active at a time. Use the `ACTIVE_PLUGIN` env variable to select the active plugin, for example:
+
+```
+ACTIVE_PLUGIN=base.eth
+```
